fix the bugs

+++ src/tendisplus/storage/repllog.cpp
@@ -486,12 +486,12 @@ uint64_t ReplLogRawV2::getTimestamp() {
 uint32_t ReplLogRawV2::getChunkId() {
     if (_val.size() < RecordValue::minSize() + ReplLogValueV2::fixedHeaderSize()) {
         INVARIANT_D(0);
-        return (uint64_t)-1;
+        return (uint32_t)-1;
     }
     auto rvHdrSize = RecordValue::decodeHdrSizeNoMeta(_val);
     if (!rvHdrSize.ok()) {
          INVARIANT_D(0);
-        return (uint64_t)-1;
+        return (uint32_t)-1;
     }
 
     return int32Decode(_val.c_str() +
